backend/requirements.txt
smartconfigparser==0.1.1
Django==1.9.6
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
# server
gunicorn==19.4.5
# test
factory-boy==2.7.0